# perf(rust, python): Rolling min/max with nulls