--- a/crates/polars-arrow/src/kernels/rolling/nulls/min_max.rs
+++ b/crates/polars-arrow/src/kernels/rolling/nulls/min_max.rs
@@ -144,7 +144,7 @@ macro_rules! minmax_window {
                     validity,
                     m: m.copied(),
                     m_idx: idx,
-                    sorted_to: sorted_to,
+                    sorted_to,
                     last_start: start,
                     last_end: end,
                     null_count,