@stinodego 2024-01-11
I am closing this PR as stale.

We have a new rolling min/max algorithm in the works, so the work here will be obsolete.